fix(cost): fail closed on partial chat pricing
